Modify NewsItem lastUpdated property * Changed from int to long

+++ app/src/main/java/com/ashleymccallum/madimalcrossing/AppDatabase.java
@@ -201,7 +201,7 @@ public class AppDatabase extends SQLiteOpenHelper {
                     cursor.getString(5),
                     cursor.getString(6),
                     cursor.getString(7),
-                    cursor.getInt(8)));
+                    cursor.getLong(8)));
         }
         return items;
     }
